refactor(challenges): use i8 for SparseChallenge coefficients and flatten unrank scan
Every shipping `SparseChallengeConfig` family caps `|coeff|` at 8 (uniform
fp128 D=32 alphabet), 2 (exact-shell), or 8 (bounded-`L1` `M`). All of those
fit comfortably in `i8` (`[-128, 127]`), so storing coefficients as `i16`
spent twice the memory bandwidth without buying anything. Switch to `i8`
throughout, and while in the area, collapse the bounded-`L1` bucket scan
from three sequential blocks (negative `a`, `a == 0`, positive `a`) to a
single `for a in -max_a..=max_a` loop.

Sampler simplifications:

* `scan_buckets_u128_unrank` returns `i8` instead of `i16` and runs one
  iteration per candidate value, with `mag = a.unsigned_abs() as usize`
  covering all three regions (including `a == 0`) uniformly. The
  loop variable is now `i8` since `max_a <= PRESET_M = 8`. Behavior is
  unchanged: same canonical visit order, same buckets, same return value.
* `sample_bounded_l1_into` takes `coeffs: &mut Vec<i8>`.
* `XofCursor::next_sign` returns `i8`.
* `sample_uniform_sparse` takes `nonzero_coeffs: &[i8]`.

Public API:

* `SparseChallenge::coeffs: Vec<i8>` (was `Vec<i16>`).
* `SparseChallengeConfig::Uniform { nonzero_coeffs: Vec<i8> }` (was
  `Vec<i16>`). Existing `vec![-1, 1]` literals at construction sites
  (`akita-config`, `akita-types`, tests) compile unchanged thanks to type
  inference.
* `SparseChallengeConfig::domain_separator_bytes` for `Uniform` now
  serializes each coefficient as 1 byte LE (was 2 bytes LE) -- a transcript
  break for the `Uniform` family. The repo carries no backwards-compat
  guarantees and no `Uniform` reference fixture pins the byte layout, so
  this is the natural shape now that the in-memory type is `i8`. The
  locked `bounded_l1_reference_vector_d32_m8_b121` fixture is unaffected
  because its serialization path doesn't depend on `nonzero_coeffs`.

Consumer fan-out:

* `akita-prover::kernels::decompose_fold_neon::sparse_mul_acc_neon` now
  takes `coeffs: &[i8]`. The body uses only `match` arms (`1`, `-1`, `2`,
  `-2`), `coeff.unsigned_abs()`, and sign comparisons, all of which work
  identically on `i8`.
* `akita-prover::backend::poly_helpers::fill_rotated_challenge` widens
  `coeff` from `i8` to `i16` at the dense-table assignment site
  (`dense[pos] = i16::from(coeff)`). The `[i16; D]` rotated table format
  is consumer-side and unchanged.
* `akita-prover::protocol::quadratic_equation`,
  `akita-challenges::challenge::{to_dense, eval_at_pows}`, and the debug
  `match coeff { 1 => ..., -1 => ... }` block in `akita-scheme::tests`
  all keep their existing `coeff as i64` / `c as i64` widening casts; the
  source type is just `i8` instead of `i16` now.

The `bounded_l1_reference_vector_d32_m8_b121` integration test continues
to pass, confirming the `i8` rewrite produces bit-identical sampled
challenges for the `BoundedL1Ball` family.

+++ crates/akita-challenges/src/challenge.rs
@@ -26,8 +26,11 @@ use akita_field::{CanonicalField, FieldCore};
 pub struct SparseChallenge {
     /// Coefficient indices (powers of `X`) where the polynomial is non-zero.
     pub positions: Vec<u32>,
-    /// Small integer coefficients at the corresponding positions.
-    pub coeffs: Vec<i16>,
+    /// Small integer coefficients at the corresponding positions. Held as
+    /// `i8` since every supported [`crate::SparseChallengeConfig`] family
+    /// caps `|coeff|` at `8` (uniform), `2` (exact-shell), or `M = 8`
+    /// (bounded-`L1`); all fit comfortably in `[-128, 127]`.
+    pub coeffs: Vec<i8>,
 }
 
 impl SparseChallenge {
