+++ crates/akita-challenges/src/sampler/xof.rs
@@ -136,7 +136,7 @@ impl XofCursor {
 
     /// Draw a uniformly random sign in `{-1, +1}`.
     #[inline]
-    pub(crate) fn next_sign(&mut self) -> i16 {
+    pub(crate) fn next_sign(&mut self) -> i8 {
         if (self.next_u8() & 1) == 0 {
             1
         } else {
